[PPAKKAD-26] Code review changes, refactored artifacts' and packages names.
