+++ examples/one-to-many/src/main/resources/akka_tracing.conf
@@ -1,4 +1,4 @@
-akka_debugging {
+akka_tracing {
   remote {
     database {
       driver = "slick.driver.SQLiteDriver$"
@@ -22,6 +22,6 @@ akka_debugging {
     }
   }
   packages = [
-    "pl.edu.agh.iet.akka_debugging.examples.actors"
+    "pl.edu.agh.iet.akka_tracing.examples.actors"
   ]
 }
