+++ examples/simple-scenario/build.sbt
@@ -1,5 +1,5 @@
 libraryDependencies ++= Seq(
   "com.typesafe.akka" %% "akka-actor" % "2.3.9",
-  "pl.edu.agh.iet" % "akka-debugging-tool-core_2.11" % "0.0.1-SNAPSHOT" % Runtime,
-  "pl.edu.agh.iet" % "akka-debugging-tool-core_2.11" % "0.0.1-SNAPSHOT"
+  "pl.edu.agh.iet" % "akka-tracing-tool-core_2.11" % "0.0.1-SNAPSHOT" % Runtime,
+  "pl.edu.agh.iet" % "akka-tracing-tool-core_2.11" % "0.0.1-SNAPSHOT"
 )
