+++ examples/simple-scenario/src/main/scala/pl/edu/agh/iet/akka_tracing/examples/actors/ThirdActor.scala
@@ -1,7 +1,7 @@
-package pl.edu.agh.iet.akka_debugging.examples.actors
+package pl.edu.agh.iet.akka_tracing.examples.actors
 
 import akka.actor.Actor
-import pl.edu.agh.iet.akka_debugging.TracedActor
+import pl.edu.agh.iet.akka_tracing.TracedActor
 
 class ThirdActor extends Actor with TracedActor {
   override def receive: Receive = {
