+++ project/Build.scala
@@ -1,4 +1,4 @@
-import pl.edu.agh.iet.akka_debugging.sbt.AkkaTracingPlugin
+import pl.edu.agh.iet.akka_tracing.sbt.AkkaTracingPlugin
 import play.sbt.PlayImport._
 import play.sbt.PlayScala
 import play.sbt.routes.RoutesKeys._
@@ -32,10 +32,10 @@ object Build extends Build {
     )
   )
 
-  lazy val root = Project("akka-debugging-tool", file("."))
+  lazy val root = Project("akka-tracing-tool", file("."))
     .settings(rootSettings: _*)
     .settings(
-      name := "akka-debugging-tool",
+      name := "akka-tracing-tool",
       publish := {},
       publishArtifact := false
     )
@@ -45,19 +45,19 @@ object Build extends Build {
       simpleScenarioExample
     )
 
-  lazy val core = Project("akka-debugging-tool-core", file("core"))
+  lazy val core = Project("akka-tracing-tool-core", file("core"))
     .settings(rootSettings: _*)
     .settings(
-      name := "akka-debugging-tool-core",
+      name := "akka-tracing-tool-core",
       libraryDependencies += "org.aspectj" % "aspectjweaver" % AspectJVersion,
       libraryDependencies += "org.aspectj" % "aspectjrt" % AspectJVersion
     )
 
-  lazy val visualization = Project("akka-debugging-tool-visualization", file("visualization"))
+  lazy val visualization = Project("akka-tracing-tool-visualization", file("visualization"))
     .enablePlugins(PlayScala)
     .settings(rootSettings: _*)
     .settings(
-      name := "akka-debugging-tool-visualization",
+      name := "akka-tracing-tool-visualization",
       libraryDependencies ++= Seq(
         jdbc,
         cache,
@@ -74,11 +74,11 @@ object Build extends Build {
     )
     .dependsOn(core)
 
-  lazy val simpleScenarioExample = Project("akka-debugging-tool-examples-simple-scenario",
+  lazy val simpleScenarioExample = Project("akka-tracing-tool-examples-simple-scenario",
     file("examples/simple-scenario"))
     .settings(rootSettings: _*)
     .settings(
-      name := "akka-debugging-tool-examples-simple-scenario",
+      name := "akka-tracing-tool-examples-simple-scenario",
       sources in(Compile, doc) := Seq.empty,
       publishArtifact in(Compile, packageDoc) := false,
       libraryDependencies ++= Seq(
@@ -92,11 +92,11 @@ object Build extends Build {
     .dependsOn(core)
     .enablePlugins(AkkaTracingPlugin)
 
-  lazy val oneToManyExample = Project("akka-debugging-tool-examples-one-to-many",
+  lazy val oneToManyExample = Project("akka-tracing-tool-examples-one-to-many",
     file("examples/one-to-many"))
     .settings(rootSettings: _*)
     .settings(
-      name := "akka-debugging-tool-examples-one-to-many",
+      name := "akka-tracing-tool-examples-one-to-many",
       sources in(Compile, doc) := Seq.empty,
       publishArtifact in(Compile, packageDoc) := false,
       libraryDependencies ++= Seq(
