+++ visualization/app/Global.scala
@@ -1,10 +1,11 @@
 import com.typesafe.config.ConfigFactory
-import pl.edu.agh.iet.akka_debugging.database.DatabaseUtils
+import pl.edu.agh.iet.akka_tracing.database.DatabaseUtils
 import play.api.libs.concurrent.Execution.Implicits._
 import play.api.{Application, GlobalSettings}
 
 import scala.concurrent.Await
 import scala.concurrent.duration._
+import scala.language.postfixOps
 
 object Global extends GlobalSettings {
   val logger = play.api.Logger.logger
