+++ visualization/app/views/index.scala.html
@@ -1,5 +1,5 @@
-@import pl.edu.agh.iet.akka_debugging.database.{CollectorDBMessage, CollectorDBMessagesRelation}
-
+@import pl.edu.agh.iet.akka_tracing.database.CollectorDBMessage
+@import pl.edu.agh.iet.akka_tracing.database.CollectorDBMessagesRelation
 @(messages: List[CollectorDBMessage])(relations: List[CollectorDBMessagesRelation])
 
 @main("Visualization tool") {
